% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contas.R
\name{somar}
\alias{somar}
\title{Function somar}
\usage{
somar(a, b)
}
\arguments{
\item{a}{first number.}

\item{b}{another number.}
}
\description{
This function is used to add two values.
}
\examples{
sum(100, 50)
Console response: 150

}
