% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contas.R
\name{subtract}
\alias{subtract}
\title{Function subtract}
\usage{
subtract(a, b)
}
\arguments{
\item{a}{first number.}

\item{b}{another number.}
}
\description{
This function is used to subtract two values.
}
\examples{
subtract(100, 50)
Console response: 50

}
